+++ frontend/src/api/profile.api.ts
@@ -1,5 +1,5 @@
 import { ProfileDto } from "../types/profile/profile.dto";
-import { instance } from "./instance.api";
+import { IApiOkResponse, instance } from "./instance.api";
 import { FollowDto } from "../types/profile/follow.dto";
 import { UnfollowDto } from "../types/profile/unfollow.dto";
 import { UpdateStatusDto } from "../types/profile/updateStatus.dto";
@@ -7,27 +7,27 @@ import { UpdateAvatarDto } from "../types/profile/updateAvatar.dto";
 import { UpdateProfileDto } from "../types/profile/updateProfile.dto";
 
 export const profileAPI = {
-    getProfile(id: number): Promise<ProfileDto> {
-        return instance.get<ProfileDto>(`/profile/${id}`).then((response) => response.data)
+    getProfile(id: number): Promise<IApiOkResponse<ProfileDto>> {
+        return instance.get<IApiOkResponse<ProfileDto>>(`/profile/${id}`).then((response) => response.data)
     },
     update(dto: UpdateProfileDto) {
         return instance.patch(`/profile/${dto.id}`, dto).then((response) => response.data)
     },
-    updateStatus(dto: UpdateStatusDto): Promise<ProfileDto> {
-        return instance.patch<ProfileDto>(`/profile/${dto.id}/status`, {status: dto.status}).then((response) => response.data)
+    updateStatus(dto: UpdateStatusDto): Promise<IApiOkResponse<ProfileDto>> {
+        return instance.patch<IApiOkResponse<ProfileDto>>(`/profile/${dto.id}/status`, {status: dto.status}).then((response) => response.data)
     },
-    updateAvatar(dto: UpdateAvatarDto): Promise<ProfileDto> {
+    updateAvatar(dto: UpdateAvatarDto): Promise<IApiOkResponse<ProfileDto>> {
         const data = new FormData();
         data.append('avatar', dto.avatar);
-        return instance.patch<ProfileDto>(`/profile/${dto.id}/avatar`, data ).then((response) => response.data)
+        return instance.patch<IApiOkResponse<ProfileDto>>(`/profile/${dto.id}/avatar`, data ).then((response) => response.data)
     },
-    follow(dto: FollowDto): Promise<ProfileDto> {
-        return instance.patch<ProfileDto>(`/profile/${dto.user_id}/follow/${dto.follow_id}`).then((response) => response.data)
+    follow(dto: FollowDto): Promise<IApiOkResponse<ProfileDto>> {
+        return instance.patch<IApiOkResponse<ProfileDto>>(`/profile/${dto.user_id}/follow/${dto.follow_id}`).then((response) => response.data)
     },
-    unfollow(dto: UnfollowDto): Promise<ProfileDto> {
-        return instance.delete<ProfileDto>(`/profile/${dto.user_id}/unfollow/${dto.unfollow_id}`).then((response) => response.data)
+    unfollow(dto: UnfollowDto): Promise<IApiOkResponse<ProfileDto>> {
+        return instance.delete<IApiOkResponse<ProfileDto>>(`/profile/${dto.user_id}/unfollow/${dto.unfollow_id}`).then((response) => response.data)
     },
-    activate(id: number): Promise<ProfileDto> {
-        return instance.get<ProfileDto>(`/profile/${id}/activate`).then((response) => response.data)
+    activate(id: number): Promise<IApiOkResponse<ProfileDto>> {
+        return instance.get<IApiOkResponse<ProfileDto>>(`/profile/${id}/activate`).then((response) => response.data)
     }
 }