+++ frontend/src/pages/Profile/ProfilePosts/Posts.tsx
@@ -6,6 +6,7 @@ import { PostDto } from "../../../types/post/post.dto";
 import { useDispatch } from "react-redux";
 import { fetchedProfile } from "../../../redux/ducks/profile/profile.slice";
 import { fetchedDelete } from "../../../redux/ducks/post/post.slice";
+import { Post } from "./Post";
 
 interface PostsInterface {
 	posts?: PostDto[];
@@ -13,25 +14,12 @@ interface PostsInterface {
 }
 
 export const Posts:React.FC<PostsInterface> = ({ posts, userId }) => {
-
-	const dispatch = useDispatch();
-
-	const onDelete = (post: PostDto) => {
-		dispatch(fetchedDelete(post.id))
-	}
-
 	return (
 		<Grid>
 			<Typography variant="h5">Posts</Typography>
 			{posts?.map((post) => {
 				return (
-					<Grid className={styles.profilePost}>
-						<Typography>{post.title}</Typography>
-						<Typography>{post.content}</Typography>
-						<Typography>Likes count: {post.likesCount}</Typography>
-						<img src={post.image} />
-						<Button onClick={() => onDelete(post)} variant="contained">Delete</Button>
-					</Grid>
+					<Post post={post}/>
 				);
 			})}
 		</Grid>